FHIR-processor/requirements.txt
azure-storage-blob>=12.16.0
azure-identity>=1.12.0
numpy<2.0.0
pandas>=1.5.0,<2.0.0
pytest>=7.0.0
python-dotenv>=0.19.0
pyyaml>=6.0 